LayoutTests/accessibility/mac/live-regions/live-region-alert-on-initial-load-expected.txt

This tests that when an alert (implicit live region) is initialized, it is announced.

Received announcement request:
AnnouncementKey: Important announcement{
    AXLanguage = en;
}
AXPriorityKey: 90
AXAnnouncementIsLiveRegionKey: true


PASS successfullyParsed is true

TEST COMPLETE
Unimportant announcement
Important announcement
